requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
black==24.4.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
exceptiongroup==1.2.2
filelock==3.15.4
flatbuffers==1.12
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.4.0
google-auth==2.32.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.9.0
Keras-Preprocessing==1.1.2
langchain==0.2.7
langchain-core==0.2.17
langchain-text-splitters==0.2.2
langsmith==0.1.85
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.10.6
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
protobuf==3.19.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pypdf==4.2.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.66.4
transformers==4.27.2
typing_extensions==4.12.2
urllib3==2.2.2
validators==0.28.3
weaviate-client==3.26.5
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
